move expo app to packages and test to see if it connects with common package
