+++ packages/expo-app/.expo-shared/assets.json
