+++ packages/expo-app/App.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import { add } from '@omega/common';
+
+export default function App() {
+	return (
+		<View style={styles.container}>
+			<Text>Open up App.js to start working on your app!</Text>
+			<Text>{add(1, 2)}</Text>
+		</View>
+	);
+}
+
+const styles = StyleSheet.create({
+	container: {
+		flex: 1,
+		backgroundColor: '#fff',
+		alignItems: 'center',
+		justifyContent: 'center',
+	},
+});
