+++ packages/expo-app/app.json
