+++ packages/expo-app/assets/icon.png
