+++ packages/expo-app/assets/splash.png
