+++ packages/expo-app/package.json
